fix synchronous functions (commandqueue.finish and waitforevents)
In e10s, the chrome priviledged code will run all user specified callbacks asynchronously, which is incompatible with webcl's synchronous functions.
It is expected that the callbacks are run when commandqueue.finish returns.

This commit fixes that (in a rather hackish way), by making sure all callbacks are called from framescript. While this change does expose more internals to the user, it does not really matter since the only thing that is accessible by the user, are the callbacks he specified himself.